+++ packages/create/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/create
 
+## 0.0.12
+
+### Patch Changes
+
+-   Updated dependencies [f8022a8]
+    -   @sap-ux/project-access@1.2.0
+    -   @sap-ux/mockserver-config-writer@0.1.0
+
+## 0.0.11
+
+### Patch Changes
+
+-   Updated dependencies [8b83e93]
+    -   @sap-ux/mockserver-config-writer@0.1.0
+
 ## 0.0.10
 
 ### Patch Changes